Check class-wide create/edit/delete flags before handling respective operations

+++ singleurlcrud/views.py
@@ -459,17 +459,22 @@ class CRUDView(PaginationMixin, ListView):
     def get(self, request, *args, **kwargs):
         try:
             if request.GET.get('o', '') == u'add':
-                if not self.check_permission('add', None, request):
+                if not self.enable_create or \
+                        not self.check_permission('add', None, request):
                     raise Http404
             elif request.GET.get('o', '') == u'edit' and request.GET.get('item'):
                 item = get_object_or_404(self.get_model(), pk=self.request.GET.get('item'))
-                if not self.check_permission('edit', item, request):
+                if not self.enable_edit or \
+                        not self.check_permission('edit', item, request):
                     raise Http404
             elif request.GET.get('o', '') == u'delete' and request.GET.get('item'):
                 item = get_object_or_404(self.get_model(), pk=self.request.GET.get('item'))
-                if not self.check_permission('delete', item, request):
+                if not self.enable_delete or \
+                        not self.check_permission('delete', item, request):
                     raise Http404
             elif request.GET.get('o', '') == u'delete_multiple' and request.GET.get('items'):
+                if not self.enable_delete:
+                    raise Http404
                 items = self.request.GET.get("items")
                 objects = self.get_queryset().filter(pk__in=items.split(","))
             else:
@@ -496,6 +501,9 @@ class CRUDView(PaginationMixin, ListView):
 
     def post_add(self, request, *args, **kwargs):
         # add a new item
+        if not self.enable_create or \
+                not self.check_permission('add', None, request):
+            raise Http404
         try:
             form = self.get_form(self.get_form_class(), data=self.request.POST)
             context_args = {}
@@ -536,6 +544,9 @@ class CRUDView(PaginationMixin, ListView):
         # edit
         try:
             item = get_object_or_404(self.get_model(), pk=self.request.GET.get('item'))
+            if not self.enable_edit or \
+                    not self.check_permission('edit', item, request):
+                raise Http404
             form = self.get_form(self.get_form_class(), instance=item, data=self.request.POST,
                         files=request.FILES)
             context_args = {}
@@ -580,11 +591,16 @@ class CRUDView(PaginationMixin, ListView):
         # delete
         if request.GET.get('items'):
             item_ids = request.GET.get('items')
+            if not self.enable_delete:
+                raise Http404
             objects = self.get_queryset().filter(pk__in=item_ids.split(","))
             objects.delete()
             msg = _('Selected %s have been deleted') % self.get_model()._meta.verbose_name_plural.title()
         else:
             item = get_object_or_404(self.get_model(), pk=self.request.GET.get('item'))
+            if not self.enable_delete or \
+                    not self.check_permission('delete', item, request):
+                raise Http404
             # verify global delete view flag and individual item deletable flag
             # (if it was specified) before doing the actual deletion.
             if not getattr(item, 'is_readonly', False) and self.item_deletable(item):
